Organize preprocessing plugins hierarchically
