+++ src/tempor/plugins/preprocessing/__init__.py
@@ -1,6 +1,7 @@
-from . import imputation, scaling
+from . import imputation, nop, scaling
 
 __all__ = [
     "imputation",
+    "nop",
     "scaling",
 ]
