+++ src/tempor/plugins/preprocessing/imputation/_base.py
@@ -1,0 +1,6 @@
+import tempor.plugins.core as plugins
+
+
+class BaseImputer(plugins.BaseTransformer):
+    def __init__(self, **params) -> None:  # pylint: disable=useless-super-delegation
+        super().__init__(**params)
