+++ src/tempor/plugins/preprocessing/imputation/static/plugin_static_imputation.py
@@ -7,16 +7,18 @@ import tempor.plugins.core as plugins
 from tempor.data import dataset
 from tempor.data.samples import StaticSamples, TimeSeriesSamples
 from tempor.plugins.core._params import CategoricalParams, Params
-from tempor.plugins.preprocessing.imputation import BaseImputer
+from tempor.plugins.preprocessing.imputation._base import BaseImputer
 
 
-@plugins.register_plugin(name="ffill", category="preprocessing.imputation")
-class FFillImputer(BaseImputer):
-    """Forward-first Time-Series Imputation.
+@plugins.register_plugin(name="static_imputation", category="preprocessing.imputation.static")
+class StaticOnlyImputer(BaseImputer):
+    """Impute the time-series using a static imputer.
 
     Args:
         static_imputer (str, optional):
-            Which imputer to use for the static data (if any). Defaults to ``"mean"``.
+            Which imputer to use for the static data (if any). Defaults to ``"ice"``.
+        temporal_imputer (str, optional):
+            Which imputer to use for the temporal data (if any). Defaults to ``"ice"``.
         random_state (int, optional):
             Random seed. Defaults to ``0``.
 
@@ -29,11 +31,11 @@ class FFillImputer(BaseImputer):
         >>> assert dataset.time_series.dataframe().isna().sum().sum() != 0
         >>>
         >>> # Load the model:
-        >>> model = plugin_loader.get("preprocessing.imputation.ffill")
+        >>> model = plugin_loader.get("preprocessing.imputation.static.static_imputation")
         >>>
         >>> # Train:
         >>> model.fit(dataset)
-        FFillImputer(...)
+        StaticOnlyImputer(...)
         >>>
         >>> # Impute:
         >>> imputed = model.transform(dataset)
@@ -42,15 +44,18 @@ class FFillImputer(BaseImputer):
     """
 
     def __init__(
-        self, static_imputer: str = "mean", random_state: int = 0, **params
+        self, static_imputer: str = "ice", temporal_imputer: str = "ice", random_state: int = 0, **params
     ) -> None:  # pylint: disable=useless-super-delegation
         super().__init__(**params)
         self.static_imputer = StaticImputers().get(static_imputer, random_state=random_state)
+        self.temporal_imputer = StaticImputers().get(temporal_imputer, random_state=random_state)
         self.random_state = random_state
 
     def _fit(self, data: dataset.Dataset, *args, **kwargs) -> Self:
         if data.static is not None:
             self.static_imputer.fit(data.static.dataframe())
+
+        self.temporal_imputer.fit(data.time_series.dataframe())
 
         return self
 
@@ -65,16 +70,17 @@ class FFillImputer(BaseImputer):
             data.static = StaticSamples.from_dataframe(imputed_static_data)
 
         # impute temporal data
-        sample_ts_index = data.time_series.sample_index()
-        imputed_ts = data.time_series.dataframe()
-        for idx in sample_ts_index:
-            imputed_ts.loc[(idx, slice(None)), :] = imputed_ts.loc[(idx, slice(None)), :].ffill()  # pyright: ignore
-            imputed_ts.loc[(idx, slice(None)), :] = imputed_ts.loc[(idx, slice(None)), :].bfill()  # pyright: ignore
-            imputed_ts.loc[(idx, slice(None)), :] = imputed_ts.loc[(idx, slice(None)), :].fillna(0.0)  # pyright: ignore
+        ts_data = data.time_series.dataframe()
+        imputed_ts_data = self.temporal_imputer.transform(ts_data)
+        imputed_ts_data.columns = ts_data.columns
+        imputed_ts_data.index = ts_data.index
 
-        data.time_series = TimeSeriesSamples.from_dataframe(imputed_ts)
+        data.time_series = TimeSeriesSamples.from_dataframe(imputed_ts_data)
         return data
 
     @staticmethod
     def hyperparameter_space(*args: Any, **kwargs: Any) -> List[Params]:  # pragma: no cover
-        return [CategoricalParams(name="static_imputer", choices=["mean", "ice", "missforest"])]
+        return [
+            CategoricalParams(name="static_imputer", choices=["mean", "ice", "missforest"]),
+            CategoricalParams(name="temporal_imputer", choices=["mean", "ice", "missforest"]),
+        ]
