+++ src/tempor/plugins/preprocessing/nop/__init__.py
@@ -4,11 +4,12 @@ from typing_extensions import Self
 
 import tempor.plugins.core as plugins
 from tempor.data import dataset
-from tempor.plugins.preprocessing.imputation import BaseImputer
+
+plugins.register_plugin_category("preprocessing.nop", plugins.BaseTransformer)
 
 
-@plugins.register_plugin(name="nop_imputer", category="preprocessing.imputation")
-class NopImputer(BaseImputer):
+@plugins.register_plugin(name="nop_transformer", category="preprocessing.nop")
+class NopTransformer(plugins.BaseTransformer):
     def __init__(self, **params) -> None:  # pylint: disable=useless-super-delegation
         super().__init__(**params)
 
@@ -26,3 +27,8 @@ class NopImputer(BaseImputer):
     @staticmethod
     def hyperparameter_space(*args, **kwargs):
         return []
+
+
+__all__ = [
+    "NopTransformer",
+]
