+++ src/tempor/plugins/preprocessing/scaling/_base.py
@@ -1,0 +1,6 @@
+import tempor.plugins.core as plugins
+
+
+class BaseScaler(plugins.BaseTransformer):
+    def __init__(self, **params) -> None:  # pylint: disable=useless-super-delegation
+        super().__init__(**params)
