+++ src/tempor/plugins/preprocessing/scaling/static/plugin_static_standard_scaler.py
@@ -1,23 +1,21 @@
 from typing import Any
 
 import pandas as pd
-from sklearn.preprocessing import MinMaxScaler
+from sklearn.preprocessing import StandardScaler
 from typing_extensions import Self
 
 import tempor.plugins.core as plugins
 from tempor.data import dataset
 from tempor.data.samples import StaticSamples
-from tempor.plugins.preprocessing.scaling import BaseScaler
+from tempor.plugins.preprocessing.scaling._base import BaseScaler
 
 
-@plugins.register_plugin(name="static_minmax_scaler", category="preprocessing.scaling")
-class StaticMinMaxScaler(BaseScaler):
+@plugins.register_plugin(name="static_standard_scaler", category="preprocessing.scaling.static")
+class StaticStandardScaler(BaseScaler):
     def __init__(self, **params) -> None:  # pylint: disable=useless-super-delegation
-        """MinMax scaling for the static data.
+        """Standard scaling for the static data.
 
-        Transform the static features by scaling each feature to a given range.
-        This estimator scales and translates each feature individually such that it is in the given range on the
-        training set, e.g. between zero and one.
+        Standardize the static features by removing the mean and scaling to unit variance.
 
         Example:
             >>> from tempor.utils.dataloaders import SineDataLoader
@@ -26,18 +24,18 @@ class StaticMinMaxScaler(BaseScaler):
             >>> dataset = SineDataLoader().load()
             >>>
             >>> # Load the model:
-            >>> model = plugin_loader.get("preprocessing.scaling.static_minmax_scaler")
+            >>> model = plugin_loader.get("preprocessing.scaling.static.static_standard_scaler")
             >>>
             >>> # Train:
             >>> model.fit(dataset)
-            StaticMinMaxScaler(...)
+            StaticStandardScaler(...)
             >>>
             >>> # Scale:
             >>> scaled = model.transform(dataset)
         """
 
         super().__init__(**params)
-        self.model = MinMaxScaler()
+        self.model = StandardScaler()
 
     def _fit(
         self,
