+++ tests/benchmarks/test_benchmark.py
@@ -29,7 +29,7 @@ def test_classifier_benchmark(data: str, request: pytest.FixtureRequest) -> None
             "pipeline1",
             Pipeline(
                 [
-                    "preprocessing.imputation.ffill",
+                    "preprocessing.imputation.temporal.ffill",
                     PREDICTOR_CLASSIFICATION,
                 ]
             )({"nn_classifier": {"n_iter": N_ITER}}),
@@ -67,7 +67,7 @@ def test_regressor_benchmark(data: str, request: pytest.FixtureRequest) -> None:
             "pipeline1",
             Pipeline(
                 [
-                    "preprocessing.imputation.ffill",
+                    "preprocessing.imputation.temporal.ffill",
                     PREDICTOR_REGRESSION,
                 ]
             )({"nn_regressor": {"n_iter": N_ITER}}),
@@ -107,7 +107,7 @@ def test_time_to_event_benchmark(
             "pipeline1",
             Pipeline(
                 [
-                    "preprocessing.imputation.ffill",
+                    "preprocessing.imputation.temporal.ffill",
                     PREDICTOR_TIME_TO_EVENT,
                 ]
             )({"dynamic_deephit": {"n_iter": N_ITER}}),
