+++ tests/plugins/preprocessing/imputation/temporal/test_bfill.py
@@ -4,8 +4,8 @@ import pytest
 
 from tempor.plugins import plugin_loader
 from tempor.plugins.preprocessing.imputation import BaseImputer
-from tempor.plugins.preprocessing.imputation.plugin_static_imputation import (
-    StaticOnlyImputer as plugin,
+from tempor.plugins.preprocessing.imputation.temporal.plugin_bfill import (
+    BFillImputer as plugin,
 )
 from tempor.utils.serialization import load, save
 
@@ -15,7 +15,7 @@ TEST_ON_DATASETS = ["sine_data_missing_small"]
 
 
 def from_api() -> BaseImputer:
-    return plugin_loader.get("preprocessing.imputation.static_imputation", **train_kwargs)
+    return plugin_loader.get("preprocessing.imputation.temporal.bfill", **train_kwargs)
 
 
 def from_module() -> BaseImputer:
@@ -25,8 +25,8 @@ def from_module() -> BaseImputer:
 @pytest.mark.parametrize("test_plugin", [from_api(), from_module()])
 def test_sanity(test_plugin: BaseImputer) -> None:
     assert test_plugin is not None
-    assert test_plugin.name == "static_imputation"
-    assert len(test_plugin.hyperparameter_space()) == 2
+    assert test_plugin.name == "bfill"
+    assert len(test_plugin.hyperparameter_space()) == 1
 
 
 @pytest.mark.parametrize(
