+++ tests/plugins/preprocessing/scaling/static/test_static_minmax_scaler.py
@@ -2,8 +2,8 @@ import pytest
 
 from tempor.plugins import plugin_loader
 from tempor.plugins.preprocessing.scaling import BaseScaler
-from tempor.plugins.preprocessing.scaling.plugin_ts_minmax_scaler import (
-    TimeSeriesMinMaxScaler as plugin,
+from tempor.plugins.preprocessing.scaling.static.plugin_static_minmax_scaler import (
+    StaticMinMaxScaler as plugin,
 )
 from tempor.utils.serialization import load, save
 
@@ -14,7 +14,7 @@ TEST_TRANSFORM_ON_DATASETS = ["sine_data_scaled_small"]
 
 
 def from_api() -> BaseScaler:
-    return plugin_loader.get("preprocessing.scaling.ts_minmax_scaler", **train_kwargs)
+    return plugin_loader.get("preprocessing.scaling.static.static_minmax_scaler", **train_kwargs)
 
 
 def from_module() -> BaseScaler:
@@ -24,7 +24,7 @@ def from_module() -> BaseScaler:
 @pytest.mark.parametrize("test_plugin", [from_api(), from_module()])
 def test_sanity(test_plugin: BaseScaler) -> None:
     assert test_plugin is not None
-    assert test_plugin.name == "ts_minmax_scaler"
+    assert test_plugin.name == "static_minmax_scaler"
     assert len(test_plugin.hyperparameter_space()) == 0
 
 
@@ -52,8 +52,8 @@ def test_fit(test_plugin: BaseScaler, data: str, request: pytest.FixtureRequest)
 @pytest.mark.parametrize("data", TEST_TRANSFORM_ON_DATASETS)
 def test_transform(test_plugin: BaseScaler, data: str, request: pytest.FixtureRequest) -> None:
     dataset = request.getfixturevalue(data)
-    assert dataset.time_series is not None  # nosec B101
-    assert (dataset.time_series.numpy() > 1.1).any()
+    assert dataset.static is not None  # nosec B101
+    assert (dataset.static.numpy() > 1.1).any()
 
     dump = save(test_plugin)
     reloaded = load(dump)
@@ -65,5 +65,5 @@ def test_transform(test_plugin: BaseScaler, data: str, request: pytest.FixtureRe
 
     output = reloaded.transform(dataset)
 
-    assert (output.time_series.numpy() < 1 + 1e-1).all()
-    assert (output.time_series.numpy() >= 0).all()
+    assert (output.static.numpy() < 1 + 1e-1).all()
+    assert (output.static.numpy() >= 0).all()
